server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Twilio
twilio==9.3.7

# Deepgram
deepgram-sdk==3.8.3

# OpenAI
openai==1.57.2

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis[hiredis]==5.2.1
aioredis==2.0.1

# Google Calendar
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0

# VIN Decoding
pyvin==1.0.1
requests==2.32.3

# Utilities
python-multipart==0.0.20
httpx==0.28.1
tenacity==9.0.0
